+++ public/index.html
@@ -48,6 +48,6 @@
       </div>
     </main>
 
-    <script src="../src/main.js"></script>
+    <script src="./bundle.js"></script>
   </body>
 </html>
