+++ src/components/const.js
@@ -1,0 +1,5 @@
+export {PROJECT__CONST};
+
+const PROJECT__CONST = {
+  taskCount: 3
+};
