+++ src/components/day-point-list.js
@@ -1,0 +1,11 @@
+export {createDayPointListTemplate};
+
+const createDayPointListTemplate = () => {
+  return (
+    `
+      <ul class="trip-events__list">
+
+      </ul>  
+    `
+  );
+};
