+++ src/components/trip-days-list.js
@@ -1,0 +1,11 @@
+export {createDaysListTemplate};
+
+const createDaysListTemplate = () => {
+  return (
+    `
+      <ul class="trip-days">
+
+      </ul>
+    `
+  );
+};
