+++ src/components/utils.js
@@ -1,0 +1,5 @@
+export {render};
+
+const render = (container, template, place) => {
+  container.insertAdjacentHTML(place, template);
+};
